# Speeding up onehotbatch by creating OneHotArray directly