--- a/src/onehot.jl
+++ b/src/onehot.jl
@@ -185,11 +185,10 @@ julia> reshape(1:15, 3, 5) * oh  # this matrix multiplication is done efficientl
 """
 onehotbatch(data, labels, default...) = _onehotbatch(data, length(labels) < 32 ? Tuple(labels) : labels, default...)
 
-# NB function barrier:
 function _onehotbatch(data, labels)
-  indices = UInt32[_findval(i, labels) for i in data]
-  if nothing in indices
-    unexpected_values = unique(data[indices .== nothing])
+  indices = UInt32[something(_findval(i, labels), 0) for i in data]
+  if 0 in indices
+    unexpected_values = unique(data[indices .== 0])
     error("Values $unexpected_values are not in labels")
   end
   return OneHotArray(indices, length(labels))
@@ -198,7 +197,7 @@ end
 function _onehotbatch(data, labels, default)
   default_index = _findval(default, labels)
   isnothing(default_index) && error("Default value $default is not in labels")
-  indices = UInt32[isnothing(_findval(i, labels)) ? default_index : _findval(i, labels) for i in data]
+  indices = UInt32[something(_findval(i, labels), default_index) for i in data]
   return OneHotArray(indices, length(labels))
 end