- Speeding up `onehotbatch` by constructing array directly

Hey there.
I noticed that the `onehotbatch` function calls the `onehot` function repeatedly, and then merges the resulting `OneHotVector`s.
This leads to the `OneHotVector` constructor being called many times, especially as data grows large. This PR changes the behaviour slightly, constructing a `OneHotArray` directly.

EDIT - Testing against master:
EDIT2 - Updated after broadcasts replaced with maps
EDIT3 - Added example benchmark cases from #1861 / #1844 
EDIT4 - Update after replacing maps with comprehension

Before:
```julia
julia> small_data = rand(MersenneTwister(0), 'a':'e', 20);
julia> small_labels = 'a':'e'
julia> big_data = rand(MersenneTwister(0), '.':'z', 10^6);
julia> big_labels = '.':'z'

julia> @btime onehotbatch($small_data, $small_labels);
  694.440 ns (5 allocations: 416 bytes)

julia> @btime onehotbatch($big_data, $big_labels);
  1.474 s (8000007 allocations: 358.58 MiB

julia> @btime onehotbatch($small_data, $big_labels);
  24.796 μs (165 allocations: 7.55 KiB)

julia> @btime onehotbatch($small_data, labels) setup=(labels=Tuple(small_labels));
  415.618 ns (3 allocations: 304 bytes)

julia> @btime onehotbatch(data, $big_labels, 'c') setup=(data=rand(MersenneTwister(0), 1:50, 10^6));
  1.657 s (8000007 allocations: 358.58 MiB)

# Example cases from #1861 / #1844
julia> const bases_dna = ['A', 'C', 'G', 'T'];
julia> dna_sequence = "CCGAGGGCTATGGTTTGGAAGTTAGAACCCTGGGGCTTCTCGCGGA";
julia> short_string = String(rand('a':'z', 10));
julia> string_labels = 'a':'z';

julia> @btime onehotbatch($dna_sequence, $bases_dna);
  793.137 ns (4 allocations: 528 bytes)

julia> @btime onehotbatch($short_string, $string_labels);
  1.512 μs (5 allocations: 480 bytes)
```

After:
```julia
julia> small_data = rand(MersenneTwister(0), 'a':'e', 20);
julia> small_labels = 'a':'e'
julia> big_data = rand(MersenneTwister(0), '.':'z', 10^6);
julia> big_labels = '.':'z'

julia> @btime onehotbatch($small_data, $small_labels);
  646.539 ns (4 allocations: 272 bytes)

julia> @btime onehotbatch($big_data, $big_labels);
  9.802 ms (3 allocations: 3.81 MiB)

julia> @btime onehotbatch($small_data, $big_labels);
  522.220 ns (2 allocations: 160 bytes)

julia> @btime onehotbatch($small_data, labels) setup=(labels=Tuple(small_labels));
  371.029 ns (2 allocations: 160 bytes)

julia> @btime onehotbatch(data, $big_labels, 'c') setup=(data=rand(MersenneTwister(0), 1:50, 10^6));
  266.723 ms (3 allocations: 3.81 MiB)

# Example cases from #1861 / #1844
julia> const bases_dna = ['A', 'C', 'G', 'T'];
julia> dna_sequence = "CCGAGGGCTATGGTTTGGAAGTTAGAACCCTGGGGCTTCTCGCGGA";
julia> short_string = String(rand('a':'z', 10));
julia> string_labels = 'a':'z';

julia> @btime onehotbatch($dna_sequence, $bases_dna);
  753.800 ns (3 allocations: 288 bytes)

julia> @btime onehotbatch($short_string, $string_labels);
  1.456 μs (4 allocations: 384 bytes)
```

For very small data/labels, the older implementation is a bit faster, but then the newer one tends to do better.

### PR Checklist

- [x] Tests are added
- [ ] Entry in NEWS.md
- [x] Documentation, if applicable