@TLipede 2022-02-22
> This looks promising. Most recent update is #1861, can you test this on those examples too?

Heya, sorry the description wasn't very clear. This is being tested against master. I'll edit the description.