@mcabbott 2022-02-22
Comparing against master is good. But what I meant is, that PR (and the issue that inspired it) have some different examples, with strings not arrays. Finding problems to benchmark that you didn't make up for the purpose occasionally reveals blind spots.