@TLipede 2022-02-23
I see what you mean now @mcabbott, added those examples now. There's a test and an error now if `default` isn't provided, but the error will only get triggered if unexpected values are encountered. Are we okay with that behaviour, or should the function always error?